moved WorldSelector
