+++ src/main/java/org/betonquest/betonquest/api/common/worldselector/package-info.java
@@ -1,4 +1,4 @@
 /**
  * This package contains classes for selecting the world to use for an event.
  */
-package org.betonquest.betonquest.quest.event.worldselector;
+package org.betonquest.betonquest.api.common.worldselector;
